requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openai>=1.0.0
PyYAML>=6.0
